* Split Script responsibilities into JRubyAdapter and Script.
