+++ test/handle_activity/stack_activity_test.rb
@@ -12,11 +12,11 @@ end
 
 test('stack depth') do |activity|
   os_offset = {13 => 1, 15 => 1, 16 => 1}[android.os.Build::VERSION::SDK_INT].to_i
-  if org.ruboto.Script.uses_platform_apk?
+  if org.ruboto.JRubyAdapter.uses_platform_apk?
     jruby_offset = {
         '0.4.7' => [0, 0, 0, 0],
         '0.4.8.dev' => [0, -1, -1, -1],
-    }[org.ruboto.Script.platform_version_name] || [0, 0, 0, 0]
+    }[org.ruboto.JRubyAdapter.platform_version_name] || [0, 0, 0, 0]
   else
     jruby_offset = {
         '1.7.0.dev' => [1, 1, 1, 1],
@@ -24,7 +24,7 @@ test('stack depth') do |activity|
         '1.7.0.preview2.dev' => [0, -1, -1, -1],
     }[org.jruby.runtime.Constants::VERSION] || [0, 0, 0, 0]
   end
-  version_message ="ANDROID: #{android.os.Build::VERSION::SDK_INT}, PLATFORM: #{org.ruboto.Script.uses_platform_apk ? org.ruboto.Script.platform_version_name : 'STANDALONE'}, JRuby: #{org.jruby.runtime.Constants::VERSION}"
+  version_message ="ANDROID: #{android.os.Build::VERSION::SDK_INT}, PLATFORM: #{org.ruboto.JRubyAdapter.uses_platform_apk ? org.ruboto.JRubyAdapter.platform_version_name : 'STANDALONE'}, JRuby: #{org.jruby.runtime.Constants::VERSION}"
   assert_equal 43 + os_offset + jruby_offset[0], activity.find_view_by_id(42).text.to_i, version_message
   assert_equal 67 + os_offset + jruby_offset[1], activity.find_view_by_id(43).text.to_i, version_message
   assert_equal 76 + os_offset + jruby_offset[2], activity.find_view_by_id(44).text.to_i, version_message
